Extract commit and discard helpers in a dedicated file
